type FiltrationOfSimplicialComplexes
The contents of FiltrationOfSimplicialComplexes.jl is now complete.
(0) The setting that “type FiltrationOfSimplicialComplexes” is based on
the constructed function “push!” is changed to being based on a new
function called “FaceBirthpush!”. The reason for the change is that an
argument of “push!” is of “type FiltrationOfSimplicialComplexes” while
the construction of “type FiltrationOfSimplicialComplexes” uses the
function “push!”, which is very confusing to me.
(1) A function “FaceBirthpush!” is written for the core construction in
type “FiltrationOfSimplicialComplexes”.
(2) The “type FiltrationOfSimplicialComplexes” is now completed.
(3) The name of the function “push!” is replaced by “Filpush!” due to
conflicts to “Base.push!” issue.
(Issue: If the name “push!” is used, “push!(::Array{Any,1},
::Array{Int,1})” would be invalid immediately.)
Moreover, its construction is now based on “FaceBirthpush!” instead of
being self-constructed.
(4) The two “Sample” functions and the function “DowkerComplex” are now
completed.